--- frontend/constants/tool_IDs.js
@@ -1,9 +1,0 @@
-export const TOOL_IDS = {
-  FLASHCARDS: '1',
-  MCQ: '0',
-};
-
-export const TOOL_NAMES = {
-  [TOOL_IDS.FLASHCARDS]: 'Flashcards',
-  [TOOL_IDS.MCQ]: 'MultipleChoiceQuiz',
-};
